+++ include/event_data.h
@@ -28,6 +28,7 @@ u8 FlagSet(u16 id);
 u8 FlagToggle(u16 id);
 u8 FlagClear(u16 id);
 bool8 FlagGet(u16 id);
+u8 getCurrentSeason(void);
 
 extern u16 gSpecialVar_0x8000;
 extern u16 gSpecialVar_0x8001;