+++ include/global.h
@@ -572,6 +572,7 @@ struct SaveBlock2
               u8 rivalName[PLAYER_NAME_LENGTH + 1];
               u16 optionsSurfOverworld:1; // whether to use the original Surf blob or dynamic blob (False to use dynamic)
               u16 optionsFontType:1;
+			  u16 optionsSeasons:1;
 }; // sizeof=0xF2C
 
 extern struct SaveBlock2 *gSaveBlock2Ptr;