+++ include/tilesets.h
@@ -3,6 +3,9 @@
 
 extern const u32 gTilesetTiles_General[];
 extern const u16 gTilesetPalettes_General[][16];
+extern const u16 gTilesetPalettes_General_Summer[][16];
+extern const u16 gTilesetPalettes_General_Autumn[][16];
+extern const u16 gTilesetPalettes_General_Winter[][16];
 
 extern const struct Tileset * const gTilesetPointer_SecretBase;
 extern const struct Tileset * const gTilesetPointer_SecretBaseRedCave;