+++ src/data/tilesets/headers.h
@@ -23,6 +23,9 @@ const struct Tileset gTileset_General =
     .metatiles = gMetatiles_General,
     .metatileAttributes = gMetatileAttributes_General,
     .callback = InitTilesetAnim_General,
+	.palettes_summer = gTilesetPalettes_General_Summer,
+    .palettes_autumn = gTilesetPalettes_General_Autumn,
+    .palettes_winter = gTilesetPalettes_General_Winter,
 };
 
 const struct Tileset gTileset_Petalburg =
@@ -35,6 +38,9 @@ const struct Tileset gTileset_Petalburg =
     .metatiles = gMetatiles_Petalburg,
     .metatileAttributes = gMetatileAttributes_Petalburg,
     .callback = InitTilesetAnim_Petalburg,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_Rustboro =
@@ -46,6 +52,9 @@ const struct Tileset gTileset_Rustboro =
     .metatiles = gMetatiles_Rustboro,
     .metatileAttributes = gMetatileAttributes_Rustboro,
     .callback = InitTilesetAnim_Rustboro,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_Dewford =
@@ -57,6 +66,9 @@ const struct Tileset gTileset_Dewford =
     .metatiles = gMetatiles_Dewford,
     .metatileAttributes = gMetatileAttributes_Dewford,
     .callback = InitTilesetAnim_Dewford,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_Slateport =
@@ -69,6 +81,9 @@ const struct Tileset gTileset_Slateport =
     .metatiles = gMetatiles_Slateport,
     .metatileAttributes = gMetatileAttributes_Slateport,
     .callback = InitTilesetAnim_Slateport,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_Mauville =
@@ -80,6 +95,9 @@ const struct Tileset gTileset_Mauville =
     .metatiles = gMetatiles_Mauville,
     .metatileAttributes = gMetatileAttributes_Mauville,
     .callback = InitTilesetAnim_Mauville,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_Lavaridge =
@@ -91,6 +109,9 @@ const struct Tileset gTileset_Lavaridge =
     .metatiles = gMetatiles_Lavaridge,
     .metatileAttributes = gMetatileAttributes_Lavaridge,
     .callback = InitTilesetAnim_Lavaridge,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_Fallarbor =
@@ -103,6 +124,9 @@ const struct Tileset gTileset_Fallarbor =
     .metatiles = gMetatiles_Fallarbor,
     .metatileAttributes = gMetatileAttributes_Fallarbor,
     .callback = InitTilesetAnim_Fallarbor,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_Fortree =
@@ -114,6 +138,9 @@ const struct Tileset gTileset_Fortree =
     .metatiles = gMetatiles_Fortree,
     .metatileAttributes = gMetatileAttributes_Fortree,
     .callback = InitTilesetAnim_Fortree,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_Lilycove =
@@ -126,6 +153,9 @@ const struct Tileset gTileset_Lilycove =
     .metatiles = gMetatiles_Lilycove,
     .metatileAttributes = gMetatileAttributes_Lilycove,
     .callback = InitTilesetAnim_Lilycove,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_Mossdeep =
@@ -138,6 +168,9 @@ const struct Tileset gTileset_Mossdeep =
     .metatiles = gMetatiles_Mossdeep,
     .metatileAttributes = gMetatileAttributes_Mossdeep,
     .callback = InitTilesetAnim_Mossdeep,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_EverGrande =
@@ -150,6 +183,9 @@ const struct Tileset gTileset_EverGrande =
     .metatileAttributes = gMetatileAttributes_EverGrande,
     .callback = InitTilesetAnim_EverGrande,
     .swapPalettes = SWAP_PAL(8),
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_Pacifidlog =
@@ -162,6 +198,9 @@ const struct Tileset gTileset_Pacifidlog =
     .metatiles = gMetatiles_Pacifidlog,
     .metatileAttributes = gMetatileAttributes_Pacifidlog,
     .callback = InitTilesetAnim_Pacifidlog,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_Sootopolis =
@@ -174,6 +213,9 @@ const struct Tileset gTileset_Sootopolis =
     .metatiles = gMetatiles_Sootopolis,
     .metatileAttributes = gMetatileAttributes_Sootopolis,
     .callback = InitTilesetAnim_Sootopolis,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_BattleFrontierOutsideWest =
@@ -185,6 +227,9 @@ const struct Tileset gTileset_BattleFrontierOutsideWest =
     .metatiles = gMetatiles_BattleFrontierOutsideWest,
     .metatileAttributes = gMetatileAttributes_BattleFrontierOutsideWest,
     .callback = InitTilesetAnim_BattleFrontierOutsideWest,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_BattleFrontierOutsideEast =
@@ -196,6 +241,9 @@ const struct Tileset gTileset_BattleFrontierOutsideEast =
     .metatiles = gMetatiles_BattleFrontierOutsideEast,
     .metatileAttributes = gMetatileAttributes_BattleFrontierOutsideEast,
     .callback = InitTilesetAnim_BattleFrontierOutsideEast,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_Building =
@@ -207,6 +255,9 @@ const struct Tileset gTileset_Building =
     .metatiles = gMetatiles_InsideBuilding,
     .metatileAttributes = gMetatileAttributes_InsideBuilding,
     .callback = InitTilesetAnim_Building,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_Shop =
@@ -218,6 +269,9 @@ const struct Tileset gTileset_Shop =
     .metatiles = gMetatiles_Shop,
     .metatileAttributes = gMetatileAttributes_Shop,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_PokemonCenter =
@@ -229,6 +283,9 @@ const struct Tileset gTileset_PokemonCenter =
     .metatiles = gMetatiles_PokemonCenter,
     .metatileAttributes = gMetatileAttributes_PokemonCenter,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_Cave =
@@ -240,6 +297,9 @@ const struct Tileset gTileset_Cave =
     .metatiles = gMetatiles_Cave,
     .metatileAttributes = gMetatileAttributes_Cave,
     .callback = InitTilesetAnim_Cave,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_PokemonSchool =
@@ -251,6 +311,9 @@ const struct Tileset gTileset_PokemonSchool =
     .metatiles = gMetatiles_PokemonSchool,
     .metatileAttributes = gMetatileAttributes_PokemonSchool,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_PokemonFanClub =
@@ -262,6 +325,9 @@ const struct Tileset gTileset_PokemonFanClub =
     .metatiles = gMetatiles_PokemonFanClub,
     .metatileAttributes = gMetatileAttributes_PokemonFanClub,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_Unused1 =
@@ -273,6 +339,9 @@ const struct Tileset gTileset_Unused1 =
     .metatiles = gMetatiles_Unused1,
     .metatileAttributes = gMetatileAttributes_Unused1,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_MeteorFalls =
@@ -284,6 +353,9 @@ const struct Tileset gTileset_MeteorFalls =
     .metatiles = gMetatiles_MeteorFalls,
     .metatileAttributes = gMetatileAttributes_MeteorFalls,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_OceanicMuseum =
@@ -295,6 +367,9 @@ const struct Tileset gTileset_OceanicMuseum =
     .metatiles = gMetatiles_OceanicMuseum,
     .metatileAttributes = gMetatileAttributes_OceanicMuseum,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_CableClub =
@@ -306,6 +381,9 @@ const struct Tileset gTileset_CableClub =
     .metatiles = gMetatiles_CableClub,
     .metatileAttributes = gMetatileAttributes_CableClub,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_SeashoreHouse =
@@ -317,6 +395,9 @@ const struct Tileset gTileset_SeashoreHouse =
     .metatiles = gMetatiles_SeashoreHouse,
     .metatileAttributes = gMetatileAttributes_SeashoreHouse,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_PrettyPetalFlowerShop =
@@ -328,6 +409,9 @@ const struct Tileset gTileset_PrettyPetalFlowerShop =
     .metatiles = gMetatiles_PrettyPetalFlowerShop,
     .metatileAttributes = gMetatileAttributes_PrettyPetalFlowerShop,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_PokemonDayCare =
@@ -339,6 +423,9 @@ const struct Tileset gTileset_PokemonDayCare =
     .metatiles = gMetatiles_PokemonDayCare,
     .metatileAttributes = gMetatileAttributes_PokemonDayCare,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_Facility =
@@ -350,6 +437,9 @@ const struct Tileset gTileset_Facility =
     .metatiles = gMetatiles_Facility,
     .metatileAttributes = gMetatileAttributes_Facility,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_BikeShop =
@@ -361,6 +451,9 @@ const struct Tileset gTileset_BikeShop =
     .metatiles = gMetatiles_BikeShop,
     .metatileAttributes = gMetatileAttributes_BikeShop,
     .callback = InitTilesetAnim_BikeShop,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_RusturfTunnel =
@@ -372,6 +465,9 @@ const struct Tileset gTileset_RusturfTunnel =
     .metatiles = gMetatiles_RusturfTunnel,
     .metatileAttributes = gMetatileAttributes_RusturfTunnel,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_SecretBaseBrownCave =
@@ -383,6 +479,9 @@ const struct Tileset gTileset_SecretBaseBrownCave =
     .metatiles = gMetatiles_SecretBaseSecondary,
     .metatileAttributes = gMetatileAttributes_SecretBaseSecondary,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_SecretBaseTree =
@@ -394,6 +493,9 @@ const struct Tileset gTileset_SecretBaseTree =
     .metatiles = gMetatiles_SecretBaseSecondary,
     .metatileAttributes = gMetatileAttributes_SecretBaseSecondary,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_SecretBaseShrub =
@@ -405,6 +507,9 @@ const struct Tileset gTileset_SecretBaseShrub =
     .metatiles = gMetatiles_SecretBaseSecondary,
     .metatileAttributes = gMetatileAttributes_SecretBaseSecondary,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_SecretBaseBlueCave =
@@ -416,6 +521,9 @@ const struct Tileset gTileset_SecretBaseBlueCave =
     .metatiles = gMetatiles_SecretBaseSecondary,
     .metatileAttributes = gMetatileAttributes_SecretBaseSecondary,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_SecretBaseYellowCave =
@@ -427,6 +535,9 @@ const struct Tileset gTileset_SecretBaseYellowCave =
     .metatiles = gMetatiles_SecretBaseSecondary,
     .metatileAttributes = gMetatileAttributes_SecretBaseSecondary,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_SecretBaseRedCave =
@@ -438,6 +549,9 @@ const struct Tileset gTileset_SecretBaseRedCave =
     .metatiles = gMetatiles_SecretBaseSecondary,
     .metatileAttributes = gMetatileAttributes_SecretBaseSecondary,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_InsideOfTruck =
@@ -449,6 +563,9 @@ const struct Tileset gTileset_InsideOfTruck =
     .metatiles = gMetatiles_InsideOfTruck,
     .metatileAttributes = gMetatileAttributes_InsideOfTruck,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_Unused2 =
@@ -460,6 +577,9 @@ const struct Tileset gTileset_Unused2 =
     .metatiles = gMetatiles_Unused2,
     .metatileAttributes = gMetatileAttributes_Unused2,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_Contest =
@@ -471,6 +591,9 @@ const struct Tileset gTileset_Contest =
     .metatiles = gMetatiles_Contest,
     .metatileAttributes = gMetatileAttributes_Contest,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_LilycoveMuseum =
@@ -482,6 +605,9 @@ const struct Tileset gTileset_LilycoveMuseum =
     .metatiles = gMetatiles_LilycoveMuseum,
     .metatileAttributes = gMetatileAttributes_LilycoveMuseum,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_BrendansMaysHouse =
@@ -493,6 +619,9 @@ const struct Tileset gTileset_BrendansMaysHouse =
     .metatiles = gMetatiles_BrendansMaysHouse,
     .metatileAttributes = gMetatileAttributes_BrendansMaysHouse,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_Lab =
@@ -504,6 +633,9 @@ const struct Tileset gTileset_Lab =
     .metatiles = gMetatiles_Lab,
     .metatileAttributes = gMetatileAttributes_Lab,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_Underwater =
@@ -515,6 +647,9 @@ const struct Tileset gTileset_Underwater =
     .metatiles = gMetatiles_Underwater,
     .metatileAttributes = gMetatileAttributes_Underwater,
     .callback = InitTilesetAnim_Underwater,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_PetalburgGym =
@@ -526,6 +661,9 @@ const struct Tileset gTileset_PetalburgGym =
     .metatiles = gMetatiles_PetalburgGym,
     .metatileAttributes = gMetatileAttributes_PetalburgGym,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_SootopolisGym =
@@ -537,6 +675,9 @@ const struct Tileset gTileset_SootopolisGym =
     .metatiles = gMetatiles_SootopolisGym,
     .metatileAttributes = gMetatileAttributes_SootopolisGym,
     .callback = InitTilesetAnim_SootopolisGym,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_GenericBuilding =
@@ -548,6 +689,9 @@ const struct Tileset gTileset_GenericBuilding =
     .metatiles = gMetatiles_GenericBuilding,
     .metatileAttributes = gMetatileAttributes_GenericBuilding,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_MauvilleGameCorner =
@@ -559,6 +703,9 @@ const struct Tileset gTileset_MauvilleGameCorner =
     .metatiles = gMetatiles_MauvilleGameCorner,
     .metatileAttributes = gMetatileAttributes_MauvilleGameCorner,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_RustboroGym =
@@ -570,6 +717,9 @@ const struct Tileset gTileset_RustboroGym =
     .metatiles = gMetatiles_RustboroGym,
     .metatileAttributes = gMetatileAttributes_RustboroGym,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_DewfordGym =
@@ -581,6 +731,9 @@ const struct Tileset gTileset_DewfordGym =
     .metatiles = gMetatiles_DewfordGym,
     .metatileAttributes = gMetatileAttributes_DewfordGym,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_MauvilleGym =
@@ -592,6 +745,9 @@ const struct Tileset gTileset_MauvilleGym =
     .metatiles = gMetatiles_MauvilleGym,
     .metatileAttributes = gMetatileAttributes_MauvilleGym,
     .callback = InitTilesetAnim_MauvilleGym,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_LavaridgeGym =
@@ -603,6 +759,9 @@ const struct Tileset gTileset_LavaridgeGym =
     .metatiles = gMetatiles_LavaridgeGym,
     .metatileAttributes = gMetatileAttributes_LavaridgeGym,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_TrickHousePuzzle =
@@ -614,6 +773,9 @@ const struct Tileset gTileset_TrickHousePuzzle =
     .metatiles = gMetatiles_TrickHousePuzzle,
     .metatileAttributes = gMetatileAttributes_TrickHousePuzzle,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_FortreeGym =
@@ -625,6 +787,9 @@ const struct Tileset gTileset_FortreeGym =
     .metatiles = gMetatiles_FortreeGym,
     .metatileAttributes = gMetatileAttributes_FortreeGym,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_MossdeepGym =
@@ -636,6 +801,9 @@ const struct Tileset gTileset_MossdeepGym =
     .metatiles = gMetatiles_MossdeepGym,
     .metatileAttributes = gMetatileAttributes_MossdeepGym,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_InsideShip =
@@ -647,6 +815,9 @@ const struct Tileset gTileset_InsideShip =
     .metatiles = gMetatiles_InsideShip,
     .metatileAttributes = gMetatileAttributes_InsideShip,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_SecretBase =
@@ -672,6 +843,9 @@ const struct Tileset gTileset_EliteFour =
     .metatiles = gMetatiles_EliteFour,
     .metatileAttributes = gMetatileAttributes_EliteFour,
     .callback = InitTilesetAnim_EliteFour,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_BattleFrontier =
@@ -683,6 +857,9 @@ const struct Tileset gTileset_BattleFrontier =
     .metatiles = gMetatiles_BattleFrontier,
     .metatileAttributes = gMetatileAttributes_BattleFrontier,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_BattlePalace =
@@ -694,6 +871,9 @@ const struct Tileset gTileset_BattlePalace =
     .metatiles = gMetatiles_BattlePalace,
     .metatileAttributes = gMetatileAttributes_BattlePalace,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_BattleDome =
@@ -705,6 +885,9 @@ const struct Tileset gTileset_BattleDome =
     .metatiles = gMetatiles_BattleDome,
     .metatileAttributes = gMetatileAttributes_BattleDome,
     .callback = InitTilesetAnim_BattleDome,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_BattleFactory =
@@ -716,6 +899,9 @@ const struct Tileset gTileset_BattleFactory =
     .metatiles = gMetatiles_BattleFactory,
     .metatileAttributes = gMetatileAttributes_BattleFactory,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_BattlePike =
@@ -727,6 +913,9 @@ const struct Tileset gTileset_BattlePike =
     .metatiles = gMetatiles_BattlePike,
     .metatileAttributes = gMetatileAttributes_BattlePike,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_BattleArena =
@@ -738,6 +927,9 @@ const struct Tileset gTileset_BattleArena =
     .metatiles = gMetatiles_BattleArena,
     .metatileAttributes = gMetatileAttributes_BattleArena,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_BattlePyramid =
@@ -749,6 +941,9 @@ const struct Tileset gTileset_BattlePyramid =
     .metatiles = gMetatiles_BattlePyramid,
     .metatileAttributes = gMetatileAttributes_BattlePyramid,
     .callback = InitTilesetAnim_BattlePyramid,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_MirageTower =
@@ -760,6 +955,9 @@ const struct Tileset gTileset_MirageTower =
     .metatiles = gMetatiles_MirageTower,
     .metatileAttributes = gMetatileAttributes_MirageTower,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_MossdeepGameCorner =
@@ -771,6 +969,9 @@ const struct Tileset gTileset_MossdeepGameCorner =
     .metatiles = gMetatiles_MossdeepGameCorner,
     .metatileAttributes = gMetatileAttributes_MossdeepGameCorner,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_IslandHarbor =
@@ -782,6 +983,9 @@ const struct Tileset gTileset_IslandHarbor =
     .metatiles = gMetatiles_IslandHarbor,
     .metatileAttributes = gMetatileAttributes_IslandHarbor,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_TrainerHill =
@@ -793,6 +997,9 @@ const struct Tileset gTileset_TrainerHill =
     .metatiles = gMetatiles_TrainerHill,
     .metatileAttributes = gMetatileAttributes_TrainerHill,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_NavelRock =
@@ -804,6 +1011,9 @@ const struct Tileset gTileset_NavelRock =
     .metatiles = gMetatiles_NavelRock,
     .metatileAttributes = gMetatileAttributes_NavelRock,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_BattleFrontierRankingHall =
@@ -815,6 +1025,9 @@ const struct Tileset gTileset_BattleFrontierRankingHall =
     .metatiles = gMetatiles_BattleFrontierRankingHall,
     .metatileAttributes = gMetatileAttributes_BattleFrontierRankingHall,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_BattleTent =
@@ -826,6 +1039,9 @@ const struct Tileset gTileset_BattleTent =
     .metatiles = gMetatiles_BattleTent,
     .metatileAttributes = gMetatileAttributes_BattleTent,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_MysteryEventsHouse =
@@ -837,6 +1053,9 @@ const struct Tileset gTileset_MysteryEventsHouse =
     .metatiles = gMetatiles_MysteryEventsHouse,
     .metatileAttributes = gMetatileAttributes_MysteryEventsHouse,
     .callback = NULL,
+    .palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };
 
 const struct Tileset gTileset_UnionRoom =
@@ -848,4 +1067,7 @@ const struct Tileset gTileset_UnionRoom =
     .metatiles = gMetatiles_UnionRoom,
     .metatileAttributes = gMetatileAttributes_UnionRoom,
     .callback = NULL,
+	.palettes_summer = NULL,
+    .palettes_autumn = NULL,
+    .palettes_winter = NULL,
 };